docker-jobs/new-dumps-listener/requirements.txt
python-dateutil~=2.8.1
pytz~=2021.1
requests~=2.25.1